Move ZeekStrings from zbuf to zeekio
