+++ zio/zeekio/writer.go
@@ -6,7 +6,7 @@ import (
 	"io"
 	"strings"
 
-	"github.com/brimsec/zq/zbuf"
+	"github.com/brimsec/zq/pkg/nano"
 	"github.com/brimsec/zq/zio"
 	"github.com/brimsec/zq/zng"
 	"github.com/brimsec/zq/zng/resolver"
@@ -51,7 +51,7 @@ func (w *Writer) Write(r *zng.Record) error {
 		}
 		w.typ = r.Type
 	}
-	values, changePrecision, err := zbuf.ZeekStrings(r, w.precision, w.format)
+	values, changePrecision, err := ZeekStrings(r, w.precision, w.format)
 	if err != nil {
 		return err
 	}
@@ -117,4 +117,41 @@ func (w *Writer) writeHeader(r *zng.Record, path string) error {
 	}
 	_, err := w.Writer.Write([]byte(s))
 	return err
+}
+
+func isHighPrecision(ts nano.Ts) bool {
+	_, ns := ts.Split()
+	return (ns/1000)*1000 != ns
+}
+
+// This returns the zeek strings for this record.  XXX We need to not use this.
+// XXX change to Pretty for output writers?... except zeek?
+func ZeekStrings(r *zng.Record, precision int, fmt zng.OutFmt) ([]string, bool, error) {
+	var ss []string
+	it := r.ZvalIter()
+	var changePrecision bool
+	for _, col := range r.Type.Columns {
+		val, _, err := it.Next()
+		if err != nil {
+			return nil, false, err
+		}
+		var field string
+		if val == nil {
+			field = "-"
+		} else if precision >= 0 && col.Type == zng.TypeTime {
+			ts, err := zng.DecodeTime(val)
+			if err != nil {
+				return nil, false, err
+			}
+			if precision == 6 && isHighPrecision(ts) {
+				precision = 9
+				changePrecision = true
+			}
+			field = string(ts.AppendFloat(nil, precision))
+		} else {
+			field = col.Type.StringOf(val, fmt, false)
+		}
+		ss = append(ss, field)
+	}
+	return ss, changePrecision, nil
 }
